+++ back/src/routers/otherRouter.js
@@ -20,6 +20,10 @@ otherRouter.post("/", async (req, res, next) => {
         const from_date = req.body.from_date ?? null;
         const to_date = req.body.to_date ?? null;
 
+        if (user_id !== req.currentUserId) {
+            throw new Error("접근권한이 없는 유저입니다.");
+        }
+
         const newOther = await otherService.createOther({
             user_id,
             title,
@@ -41,7 +45,7 @@ otherRouter.post("/", async (req, res, next) => {
 otherRouter.get("/:id", async (req, res, next) => {
     try {
         const other_id = req.params.id;
-        const other = await otherService.getOther({ other_id });
+        const other = await otherService.getOtherById({ other_id });
 
         if (other.errorMessage) {
             throw new Error(other.errorMessage);
@@ -72,8 +76,14 @@ otherRouter.get("/list/:user_id", async (req, res, next) => {
 
 otherRouter.put("/:id", async (req, res, next) => {
     try {
-        const user_id = req.currentUserId;
         const other_id = req.params.id;
+        const other = await otherService.getOtherById({ other_id });
+
+        if (other.user_id !== req.currentUserId) {
+            throw new Error("접근권한이 없는 유저입니다.");
+        }
+
+        const user_id = req.currentUserId;
         const title = req.body.title ?? null;
         const description = req.body.description ?? null;
         const from_date = req.body.from_date ?? null;
@@ -100,6 +110,12 @@ otherRouter.put("/:id", async (req, res, next) => {
 otherRouter.delete("/:id", async (req, res, next) => {
     try {
         const other_id = req.params.id;
+        const other = await otherService.getOtherById({ other_id });
+
+        if (other.user_id !== req.currentUserId) {
+            throw new Error("접근권한이 없는 유저입니다.");
+        }
+
         const deletedOther = await otherService.deleteOther({ other_id });
 
         if (deletedOther.errorMessage) {
